Preliminary update of the `load` library.
Solves a conflict with the other JSON-based format: JSON Schema.
Still a lot of work to do though.
