+++ load/src/lib.rs
@@ -13,10 +13,14 @@ use treeldr::{
 	vocab::{GraphLabel, Object},
 	BlankIdIndex, Id, IriIndex,
 };
-use treeldr_syntax as syntax;
 
+mod error;
 mod source;
+mod document;
+
+pub use error::*;
 pub use source::*;
+pub use document::Document;
 
 pub use treeldr::reporting;
 pub type BuildContext = treeldr_build::Context<source::Metadata>;
@@ -48,470 +52,6 @@ pub fn build_all<V: VocabularyMut<Iri = IriIndex, BlankId = BlankIdIndex>>(
 		.map_err(BuildAllError::Build)
 }
 
-#[derive(Error, Debug)]
-pub enum LoadError {
-	#[error("unsupported MIME type `{0}`")]
-	UnsupportedMimeType(source::MimeType),
-
-	#[error("unrecognized format for file `{0}`")]
-	UnrecognizedFormat(PathBuf),
-
-	#[error("unable to read file `{0}`: {1}")]
-	UnableToRead(PathBuf, std::io::Error),
-
-	#[error("parse error")]
-	Parsing(#[from] Meta<ParseError, source::Metadata>),
-}
-
-pub struct TreeLdrDocument {
-	doc: syntax::Document<source::Metadata>,
-	local_context: syntax::build::LocalContext<source::Metadata>,
-}
-
-impl TreeLdrDocument {
-	fn declare<V: VocabularyMut<Iri = IriIndex, BlankId = BlankIdIndex>>(
-		&mut self,
-		context: &mut BuildContext,
-		vocabulary: &mut V,
-		generator: &mut impl Generator<V>,
-	) -> Result<(), syntax::build::Error<source::Metadata>> {
-		use treeldr_build::Document;
-		self.doc
-			.declare(&mut self.local_context, context, vocabulary, generator)
-	}
-
-	fn build<V: VocabularyMut<Iri = IriIndex, BlankId = BlankIdIndex>>(
-		mut self,
-		context: &mut BuildContext,
-		vocabulary: &mut V,
-		generator: &mut impl Generator<V>,
-	) -> Result<(), syntax::build::Error<source::Metadata>> {
-		use treeldr_build::Document;
-		self.doc
-			.define(&mut self.local_context, context, vocabulary, generator)
-	}
-}
-
-#[derive(Debug)]
-pub enum BuildAllError {
-	Declaration(LangError),
-	Link(LangError),
-	Build(treeldr_build::Error<source::Metadata>),
-}
-
-impl treeldr::reporting::DiagnoseWithVocabulary<source::Metadata> for BuildAllError {
-	fn message(
-		&self,
-		vocabulary: &impl Vocabulary<Iri = IriIndex, BlankId = BlankIdIndex>,
-	) -> String {
-		match self {
-			Self::Declaration(e) => e.message(vocabulary),
-			Self::Link(e) => e.message(vocabulary),
-			Self::Build(e) => e.message(vocabulary),
-		}
-	}
-
-	fn labels(
-		&self,
-		vocabulary: &impl Vocabulary<Iri = IriIndex, BlankId = BlankIdIndex>,
-	) -> Vec<codespan_reporting::diagnostic::Label<source::FileId>> {
-		match self {
-			Self::Declaration(e) => e.labels(vocabulary),
-			Self::Link(e) => e.labels(vocabulary),
-			Self::Build(e) => e.labels(vocabulary),
-		}
-	}
-
-	fn notes(
-		&self,
-		vocabulary: &impl Vocabulary<Iri = IriIndex, BlankId = BlankIdIndex>,
-	) -> Vec<String> {
-		match self {
-			Self::Declaration(e) => e.notes(vocabulary),
-			Self::Link(e) => e.notes(vocabulary),
-			Self::Build(e) => e.notes(vocabulary),
-		}
-	}
-}
-
-#[derive(Debug)]
-pub enum LangError {
-	TreeLdr(syntax::build::Error<source::Metadata>),
-	NQuads(treeldr_build::Error<source::Metadata>),
-	#[cfg(feature = "turtle")]
-	Turtle(turtle_syntax::build::MetaError<source::Metadata>),
-	#[cfg(feature = "json-schema")]
-	JsonSchema(treeldr_json_schema::import::Error<source::Metadata>),
-}
-
-impl From<syntax::build::Error<source::Metadata>> for LangError {
-	fn from(e: syntax::build::Error<source::Metadata>) -> Self {
-		Self::TreeLdr(e)
-	}
-}
-
-#[cfg(feature = "turtle")]
-impl From<turtle_syntax::build::MetaError<source::Metadata>> for LangError {
-	fn from(e: turtle_syntax::build::MetaError<source::Metadata>) -> Self {
-		Self::Turtle(e)
-	}
-}
-
-#[cfg(feature = "json-schema")]
-impl From<treeldr_json_schema::import::Error<source::Metadata>> for LangError {
-	fn from(e: treeldr_json_schema::import::Error<source::Metadata>) -> Self {
-		Self::JsonSchema(e)
-	}
-}
-
-impl treeldr::reporting::DiagnoseWithVocabulary<source::Metadata> for LangError {
-	fn message(
-		&self,
-		vocabulary: &impl Vocabulary<Iri = IriIndex, BlankId = BlankIdIndex>,
-	) -> String {
-		match self {
-			Self::TreeLdr(e) => e.message(vocabulary),
-			Self::NQuads(e) => e.message(vocabulary),
-			#[cfg(feature = "turtle")]
-			Self::Turtle(e) => e.to_string(),
-			#[cfg(feature = "json-schema")]
-			Self::JsonSchema(e) => e.message(vocabulary),
-		}
-	}
-
-	fn labels(
-		&self,
-		vocabulary: &impl Vocabulary<Iri = IriIndex, BlankId = BlankIdIndex>,
-	) -> Vec<codespan_reporting::diagnostic::Label<source::FileId>> {
-		match self {
-			Self::TreeLdr(e) => e.labels(vocabulary),
-			Self::NQuads(e) => e.labels(vocabulary),
-			#[cfg(feature = "turtle")]
-			Self::Turtle(_) => Vec::new(),
-			#[cfg(feature = "json-schema")]
-			Self::JsonSchema(e) => e.labels(vocabulary),
-		}
-	}
-
-	fn notes(
-		&self,
-		vocabulary: &impl Vocabulary<Iri = IriIndex, BlankId = BlankIdIndex>,
-	) -> Vec<String> {
-		match self {
-			Self::TreeLdr(e) => e.notes(vocabulary),
-			Self::NQuads(e) => e.notes(vocabulary),
-			#[cfg(feature = "turtle")]
-			Self::Turtle(_) => Vec::new(),
-			#[cfg(feature = "json-schema")]
-			Self::JsonSchema(e) => e.notes(vocabulary),
-		}
-	}
-}
-
-pub enum Document {
-	TreeLdr(Box<TreeLdrDocument>),
-
-	NQuads(nquads_syntax::Document<source::Metadata>),
-
-	#[cfg(feature = "turtle")]
-	Turtle(turtle_syntax::Document<source::Metadata>),
-
-	#[cfg(feature = "json-schema")]
-	JsonSchema(Box<treeldr_json_schema::Schema>),
-}
-
-pub enum DeclaredDocument {
-	TreeLdr(Box<TreeLdrDocument>),
-
-	NQuads(Dataset),
-
-	#[cfg(feature = "turtle")]
-	Turtle(Dataset),
-
-	#[cfg(feature = "json-schema")]
-	JsonSchema(Box<treeldr_json_schema::Schema>),
-}
-
-impl Document {
-	/// Load the document located at the given `path`.
-	pub fn load<'f, P>(
-		files: &'f mut source::Files<P>,
-		filename: &Path,
-	) -> Result<(Self, source::FileId), LoadError>
-	where
-		P: Clone + Eq + Hash + DisplayPath<'f> + for<'a> From<&'a Path>,
-	{
-		match files.load(&filename, None, None) {
-			Ok(file_id) => {
-				let document = match files.get(file_id).unwrap().mime_type() {
-					Some(type_) => Self::from_file_id(files, file_id, type_)?,
-					None => return Err(LoadError::UnrecognizedFormat(filename.to_owned())),
-				};
-
-				Ok((document, file_id))
-			}
-			Err(e) => Err(LoadError::UnableToRead(filename.to_owned(), e)),
-		}
-	}
-
-	/// Lead a document from its content.
-	pub fn load_content<'f, P>(
-		files: &'f mut source::Files<P>,
-		source: P,
-		content: String,
-		base_iri: Option<IriBuf>,
-		type_: MimeType,
-	) -> Result<(Self, source::FileId), LoadError>
-	where
-		P: Clone + Eq + Hash + DisplayPath<'f>,
-	{
-		let file_id = files.load_content(source, base_iri, Some(type_), content);
-		let document = Self::from_file_id(files, file_id, type_)?;
-		Ok((document, file_id))
-	}
-
-	fn from_file_id<'f, P>(
-		files: &'f mut source::Files<P>,
-		file_id: FileId,
-		type_: MimeType,
-	) -> Result<Self, LoadError>
-	where
-		P: DisplayPath<'f>,
-	{
-		match type_ {
-			MimeType::TreeLdr => Ok(Self::TreeLdr(Box::new(import_treeldr(files, file_id)?))),
-			MimeType::NQuads => Ok(Self::NQuads(import_nquads(files, file_id)?)),
-			#[cfg(feature = "turtle")]
-			MimeType::Turtle => Ok(Self::Turtle(import_turtle(files, file_id)?)),
-			#[cfg(feature = "json-schema")]
-			MimeType::JsonSchema => Ok(Self::JsonSchema(Box::new(import_json_schema(
-				files, file_id,
-			)))),
-			#[allow(unreachable_patterns)]
-			mime_type => Err(LoadError::UnsupportedMimeType(mime_type)),
-		}
-	}
-
-	fn declare<V: VocabularyMut<Iri = IriIndex, BlankId = BlankIdIndex>>(
-		self,
-		context: &mut BuildContext,
-		vocabulary: &mut V,
-		generator: &mut impl Generator<V>,
-	) -> Result<DeclaredDocument, LangError> {
-		match self {
-			Self::TreeLdr(mut d) => {
-				d.declare(context, vocabulary, generator)?;
-				Ok(DeclaredDocument::TreeLdr(d))
-			}
-			Self::NQuads(d) => {
-				let dataset: Dataset = d
-					.into_iter()
-					.map(|Meta(quad, meta)| {
-						Meta(
-							quad.insert_into(vocabulary)
-								.map_predicate(|Meta(p, m)| Meta(Id::Iri(p), m)),
-							meta,
-						)
-					})
-					.collect();
-
-				use treeldr_build::Document;
-				dataset
-					.declare(&mut (), context, vocabulary, generator)
-					.map_err(LangError::NQuads)?;
-				Ok(DeclaredDocument::NQuads(dataset))
-			}
-			#[cfg(feature = "turtle")]
-			Self::Turtle(d) => {
-				let dataset: Dataset = d
-					.build_triples_with(None, vocabulary, &mut *generator)?
-					.into_iter()
-					.map(|Meta(triple, meta)| {
-						Meta(
-							triple
-								.map_predicate(|Meta(p, m)| Meta(Id::Iri(p), m))
-								.into_quad(None),
-							meta,
-						)
-					})
-					.collect();
-
-				use treeldr_build::Document;
-				dataset
-					.declare(&mut (), context, vocabulary, generator)
-					.map_err(LangError::NQuads)?;
-				Ok(DeclaredDocument::NQuads(dataset))
-			}
-			#[cfg(feature = "json-schema")]
-			Self::JsonSchema(s) => {
-				treeldr_json_schema::import_schema(&s, None, context, vocabulary, generator)?;
-				Ok(DeclaredDocument::JsonSchema(s))
-			}
-		}
-	}
-}
-
-impl DeclaredDocument {
-	fn build<V: VocabularyMut<Iri = IriIndex, BlankId = BlankIdIndex>>(
-		self,
-		context: &mut BuildContext,
-		vocabulary: &mut V,
-		generator: &mut impl Generator<V>,
-	) -> Result<(), LangError> {
-		match self {
-			Self::TreeLdr(d) => {
-				d.build(context, vocabulary, generator)?;
-				Ok(())
-			}
-			Self::NQuads(d) => {
-				use treeldr_build::Document;
-				d.define(&mut (), context, vocabulary, generator)
-					.map_err(LangError::NQuads)?;
-				Ok(())
-			}
-			#[cfg(feature = "turtle")]
-			Self::Turtle(d) => {
-				use treeldr_build::Document;
-				d.define(&mut (), context, vocabulary, generator)
-					.map_err(LangError::NQuads)?;
-				Ok(())
-			}
-			#[cfg(feature = "json-schema")]
-			Self::JsonSchema(_) => Ok(()),
-		}
-	}
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum ParseError {
-	#[error("TreeLDR syntax error")]
-	TreeLdr(Box<treeldr_syntax::parsing::Error<treeldr_syntax::lexing::Error>>),
-
-	#[error("Turtle syntax error")]
-	Turtle(Box<turtle_syntax::parsing::Error<turtle_syntax::lexing::Error>>),
-
-	#[error("N-Quads syntax error")]
-	NQuads(Box<nquads_syntax::parsing::Error<nquads_syntax::lexing::Error>>),
-}
-
-impl ParseError {
-	pub fn diagnostic(
-		self,
-		meta: source::Metadata,
-	) -> codespan_reporting::diagnostic::Diagnostic<source::FileId> {
-		match self {
-			Self::TreeLdr(e) => Meta(e, meta).diagnostic(),
-			Self::NQuads(e) => codespan_reporting::diagnostic::Diagnostic::error()
-				.with_message("parse error")
-				.with_labels(vec![meta
-					.optional_location()
-					.unwrap()
-					.as_primary_label()
-					.with_message(e.to_string())]),
-			Self::Turtle(e) => codespan_reporting::diagnostic::Diagnostic::error()
-				.with_message("parse error")
-				.with_labels(vec![meta
-					.optional_location()
-					.unwrap()
-					.as_primary_label()
-					.with_message(e.to_string())]),
-		}
-	}
-
-	pub fn display_and_exit<'a, P: source::DisplayPath<'a>>(
-		self,
-		files: &'a source::Files<P>,
-		meta: source::Metadata,
-	) {
-		let diagnostic = self.diagnostic(meta);
-		let writer = StandardStream::stderr(ColorChoice::Always);
-		let config = codespan_reporting::term::Config::default();
-		term::emit(&mut writer.lock(), &config, files, &diagnostic).expect("diagnostic failed");
-		std::process::exit(1);
-	}
-}
-
-/// Import a TreeLDR file.
-pub fn import_treeldr<'f, P>(
-	files: &'f source::Files<P>,
-	source_id: source::FileId,
-) -> Result<TreeLdrDocument, Meta<ParseError, source::Metadata>>
-where
-	P: DisplayPath<'f>,
-{
-	use syntax::Parse;
-	let file = files.get(source_id).unwrap();
-
-	log::debug!("ready for parsing.");
-	match syntax::Document::parse_str(file.buffer().as_str(), |span| {
-		source::Metadata::Extern(Location::new(source_id, span))
-	}) {
-		Ok(doc) => {
-			log::debug!("parsing succeeded.");
-			Ok(TreeLdrDocument {
-				doc: doc.into_value(),
-				local_context: syntax::build::LocalContext::new(
-					file.base_iri().map(|iri| iri.into()),
-				),
-			})
-		}
-		Err(e) => Err(e.map(ParseError::TreeLdr)),
-	}
-}
-
 /// RDF dataset.
 pub type Dataset =
 	grdf::meta::BTreeDataset<Id, Id, Object<source::Metadata>, GraphLabel, source::Metadata>;
-
-/// Import a N-Quads file.
-pub fn import_nquads<'f, P>(
-	files: &'f source::Files<P>,
-	source_id: source::FileId,
-) -> Result<nquads_syntax::Document<source::Metadata>, Meta<ParseError, source::Metadata>>
-where
-	P: DisplayPath<'f>,
-{
-	use nquads_syntax::Parse;
-	let file = files.get(source_id).unwrap();
-	match nquads_syntax::Document::parse_str(file.buffer().as_str(), |span| {
-		source::Metadata::Extern(Location::new(source_id, span))
-	}) {
-		Ok(Meta(doc, _)) => {
-			log::debug!("parsing succeeded.");
-			Ok(doc)
-		}
-		Err(Meta(e, meta)) => Err(Meta(ParseError::NQuads(e), meta)),
-	}
-}
-
-/// Import a RDF Turtle file.
-#[cfg(feature = "turtle")]
-pub fn import_turtle<'f, P>(
-	files: &'f source::Files<P>,
-	source_id: source::FileId,
-) -> Result<turtle_syntax::Document<source::Metadata>, Meta<ParseError, source::Metadata>>
-where
-	P: DisplayPath<'f>,
-{
-	use turtle_syntax::Parse;
-	let file = files.get(source_id).unwrap();
-	match turtle_syntax::Document::parse_str(file.buffer().as_str(), |span| {
-		source::Metadata::Extern(Location::new(source_id, span))
-	}) {
-		Ok(Meta(doc, _)) => {
-			log::debug!("parsing succeeded.");
-			Ok(doc)
-		}
-		Err(Meta(e, meta)) => Err(Meta(ParseError::Turtle(e), meta)),
-	}
-}
-
-#[cfg(feature = "json-schema")]
-pub fn import_json_schema<P>(
-	files: &source::Files<P>,
-	source_id: source::FileId,
-) -> treeldr_json_schema::Schema {
-	let file = files.get(source_id).unwrap();
-	let json: serde_json::Value = serde_json::from_str(file.buffer()).expect("invalid JSON");
-	treeldr_json_schema::Schema::try_from(json).expect("invalid JSON Schema")
-}
