+++ modules/json-schema/Cargo.toml
@@ -11,7 +11,7 @@ log.workspace = true
 iref.workspace = true
 clap = { workspace = true, features = ["derive"] }
 derivative.workspace = true
-serde_json.workspace = true
+json-syntax.workspace = true
 thiserror.workspace = true
 
 # For the import function.
